+++ src/runner/prettyStack.js
@@ -12,6 +12,12 @@ const themes = {
 
 function prettyStack({ error, theme = themes.herbs }) {
 
+    if (!(error instanceof Error)) {
+        /* eslint-disable no-console */
+        console.info(`\n       `, chalk.ansi256(theme['main'])('Error:'), chalk.ansi256(theme['message'])(chalk.bgAnsi256(theme['backgroundMessage'])(error)), `\n`)
+        return
+    }
+
     // Replace the error message with a colorized version
     let messageStack = error.stack.replace(error.message, chalk.ansi256(theme['message'])(chalk.bgAnsi256(theme['backgroundMessage'])(error.message)))